requirements.txt
africastalking==2.0.2
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
boto3==1.34.34
botocore==1.34.162
celery==5.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==42.0.5
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==23.5
django-timezone-field==7.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.11
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
phonenumbers==8.13.37
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
pyotp==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
responses==0.25.8
s3transfer==0.10.4
schema==0.7.8
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.0
django-storages
